Re Connecting React to Rails API - Status - some success, some probs...
Success: I can get test message to render from Rails in JSON format, when
I visit http://localhost:3001/api/getMessage in the browser.

Problem: React keeps making the fetch request to port 3000, when
Rails is running on 3001.  So something is off.  Procfile looks ok
and changed it to what I had in Browseum, but still facing
this problem.  To be debugged further."